gtkrc theme modifications to make things render correctly.
Added commented stuff and links at the beginning for future reference, when coming back to this topic of gtk themes.
We are now using the xamarin engine by default, allowing for GtkCombos to give feedback when they have focus.
Had to set the priority for GtkAccelLabels to highest, to make white/bold properties take effect.
Removed padding from GtkMenuBar to get rid of uncomfortable pixel-wide background on the edges of the menu bars, perceptible when clicking menus on the etremes of the bar.